pagewalker/docs/requirements-manual.txt
# Used only to regenerate DOCX/PDF from PAGEWALKER_OPERATIONS_MANUAL.md
python-docx>=1.1.0
reportlab>=4.0.0
